+++ RELEASE_NOTES.md
@@ -2,12 +2,15 @@ Release Notes
 =============
 
 ## 1.7.27
+* Action Groups: Adds support for action groups to perform actions when alerts fire.
 * Functions: AzureWebJobsDashboard setting is only added for a V1 Function app
+* Functions: Add the correct netFrameworkVersion to the template based on the runtime version
 
 ## 1.7.26
 * Container Apps: Add support for Dapr components
 * Route Servers: Support 'depends_on' for routeServerBGPConnection.
 * WebApps: When creating keyvault secrets for web app appsettings replace characters unsupported in keyvault secret names with `-`
+* Functions: Add the correct netFrameworkVersion to the template based on the runtime version
 
 ## 1.7.25
 * Container Apps: Fix storage queue KEDA scaling rule authentication